test(rpc): consolidate playback fixtures
